refactor(collapsible-drawer): fix typo in component name
