+++ components/index.js
@@ -16,7 +16,7 @@ export { default as CartSidebar } from './cart-sidebar';
 export { default as CartSummary } from './cart-summary';
 export { default as CartTotals } from './cart-totals';
 export { default as CheckoutItemCard } from './checkout-item-card';
-export { default as CollapsableDrawer } from './collapsable-drawer';
+export { default as CollapsibleDrawer } from './collapsible-drawer';
 export { default as ContentWrapper } from './content-wrapper';
 export { default as Drawer } from './drawer';
 export { default as FeatureHighlight } from './feature-highlight';
